Update the purpose and description
I used a bootstrap file in place of this python script to load the apps, packages and mods using Linux commands. The initial purpose was to run this python script to load each command line in the list to the terminal. Instead of taking this approach I decided to use a bootstrap file in place of a python script. In its current state the script will read each command in the tuple and echo's that command to the terminal display. Then read the full list of install packages to the terminal. I did not yet add a function to call each command as it is muh simpler to use a bootstrap file or a bash file to accomplish this auto load task.